move imports, util funcs, const and structs definition from SourceUnit to Contract writer
