+++ src/cairoWriter/writers/cairoContractWriter.ts
@@ -1,6 +1,11 @@
-import { ASTWriter, ContractKind, SrcDesc } from 'solc-typed-ast';
+import { ASTWriter, ContractKind, SourceUnit, SrcDesc } from 'solc-typed-ast';
 import { isExternallyVisible } from '../../utils/utils';
-import { CairoContract } from '../../ast/cairoNodes';
+import {
+  CairoContract,
+  CairoGeneratedFunctionDefinition,
+  CairoImportFunctionDefinition,
+  FunctionStubKind,
+} from '../../ast/cairoNodes';
 import { TEMP_INTERFACE_SUFFIX } from '../../utils/nameModifiers';
 import { CairoASTNodeWriter } from '../base';
 import {
@@ -10,6 +15,8 @@ import {
   INDENT,
 } from '../utils';
 import { interfaceNameMappings } from './sourceUnitWriter';
+import assert from 'assert';
+import { getStructs } from '../../freeStructWritter';
 
 export class CairoContractWriter extends CairoASTNodeWriter {
   writeInner(node: CairoContract, writer: ASTWriter): SrcDesc {
@@ -56,7 +63,7 @@ export class CairoContractWriter extends CairoASTNodeWriter {
       .filter((func) => !isExternallyVisible(func))
       .map((func) => writer.write(func));
 
-    const events = node.vEvents.map((value) => writer.write(value));
+    const events = node.vEvents.map((value) => writer.write(value)).join('\n\n');
 
     const body = [...variables, ...enums, ...otherFunctions]
       .join('\n\n')
@@ -114,21 +121,88 @@ export class CairoContractWriter extends CairoASTNodeWriter {
             '}',
           ]
         : []),
-    ].join('\n');
+    ]
+      .map((l) => (l.length > 0 ? INDENT + l : l))
+      .join('\n');
+
+    // Data about imports used, util funcs, constants and structs definition are stored in the Source Unit node
+    // but should be printed inside the contract module
+    const sourceUnit = node.parent;
+    assert(sourceUnit instanceof SourceUnit, 'Contract node parent should be a Source Unit node.');
+
+    // Only constants generated by `newToDeploy` exist at this stage
+    const constants = sourceUnit.vVariables
+      .map((v) => {
+        assert(v.vValue !== undefined, 'Constants cannot be unanssigned');
+        return [`// ${v.documentation}`, `const ${v.name} = ${writer.write(v.vValue)};`].join('\n');
+      })
+      .join('\n');
+
+    const freeStructs = getStructs(sourceUnit, this.ast);
+    const structs = [...freeStructs, ...sourceUnit.vStructs, ...(node?.vStructs || [])]
+      .map((v) => writer.write(v))
+      .join('\n\n');
+
+    const importFunctions = sourceUnit.vFunctions.filter(
+      (f): f is CairoImportFunctionDefinition => f instanceof CairoImportFunctionDefinition,
+    );
+    const generatedFunctions = sourceUnit.vFunctions.filter(
+      (f): f is CairoGeneratedFunctionDefinition => f instanceof CairoGeneratedFunctionDefinition,
+    );
+    const otherSourceUnitFunctions = sourceUnit.vFunctions.filter(
+      (f) =>
+        !(f instanceof CairoGeneratedFunctionDefinition) &&
+        !(f instanceof CairoImportFunctionDefinition),
+    );
+
+    const writtenImportFuncs = getGroupedImports(
+      importFunctions
+        .sort((funcA, funcB) =>
+          `${funcA.path}.${funcA.name}`.localeCompare(`${funcB.path}.${funcB.name}`),
+        )
+        .filter((func, index, importFuncs) => func.name !== importFuncs[index - 1]?.name),
+    ).reduce((writtenImports, importFunc) => `${writtenImports}\n${importFunc}`, '');
+
+    const writtenGeneratedFuncs = generatedFunctions
+      .sort((funcA, funcB) => funcA.name.localeCompare(funcB.name))
+      .sort((funcA, funcB) => {
+        const stubA = funcA.functionStubKind;
+        const stubB = funcB.functionStubKind;
+        if (stubA === stubB) return 0;
+        if (stubA === FunctionStubKind.StructDefStub) return -1;
+        if (stubA === FunctionStubKind.StorageDefStub) return -1;
+        return 1;
+      })
+      .filter((func, index, genFuncs) => func.name !== genFuncs[index - 1]?.name)
+      .map((func) => writer.write(func))
+      .join('\n\n');
+
+    const writtenOtherSourceUnitFunctions = otherSourceUnitFunctions
+      .map((func) => writer.write(func))
+      .join('\n\n');
+
+    const contractHeader = '#[contract] \n' + `mod ${node.name} {`;
 
     return [
       [
+        contractHeader,
         documentation,
-        ...events,
-        `namespace ${node.name}{\n\n${body}\n\n}`,
-        outsideNamespaceBody,
+        writtenImportFuncs,
+        constants,
         storageCode,
+        events,
+        structs,
+        body,
+        outsideNamespaceBody,
+        writtenGeneratedFuncs,
+        writtenOtherSourceUnitFunctions,
+        `}`,
       ].join('\n\n'),
     ];
   }
 
   writeWhole(node: CairoContract, writer: ASTWriter): SrcDesc {
-    return [`// Contract Def ${node.name}\n\n${this.writeInner(node, writer)}`];
+    return [`${this.writeInner(node, writer)}`];
   }
 
   private writeContractInterface(node: CairoContract, writer: ASTWriter): SrcDesc {
@@ -167,4 +241,18 @@ export class CairoContractWriter extends CairoASTNodeWriter {
       ].join('\n'),
     ];
   }
+}
+
+function getGroupedImports(imports: CairoImportFunctionDefinition[]): string[] {
+  const processedImports: string[] = [];
+  imports.reduce((functionNames: string[], importNode, index) => {
+    functionNames.push(importNode.name);
+    if (importNode.path !== imports[index + 1]?.path) {
+      // TODO: multiple imports are not avaible in cairo1
+      processedImports.push(`use ${importNode.path}::${functionNames.join(', ')}`);
+      functionNames = [];
+    }
+    return functionNames;
+  }, []);
+  return processedImports;
 }
