+++ src/cairoWriter/writers/sourceUnitWriter.ts
@@ -1,11 +1,5 @@
 import assert from 'assert';
 import { ASTWriter, ContractKind, SourceUnit, SrcDesc } from 'solc-typed-ast';
-import {
-  CairoImportFunctionDefinition,
-  CairoGeneratedFunctionDefinition,
-  FunctionStubKind,
-} from '../../ast/cairoNodes';
-import { getStructsAndRemappings } from '../../freeStructWritter';
 import { removeExcessNewlines } from '../../utils/formatting';
 import { TEMP_INTERFACE_SUFFIX } from '../../utils/nameModifiers';
 import { CairoASTNodeWriter } from '../base';
@@ -27,75 +21,13 @@ export class SourceUnitWriter extends CairoASTNodeWriter {
         ? node.vContracts.filter((cd) => !cd.name.endsWith(TEMP_INTERFACE_SUFFIX))
         : node.vContracts;
 
-    assert(mainContract_.length <= 1);
-    const [mainContract] = mainContract_;
+    assert(mainContract_.length <= 1, 'Every SourceUnit should only define a single contract');
 
-    const [freeStructs, freeStructRemappings_] = mainContract
-      ? getStructsAndRemappings(node, this.ast)
-      : [[], new Map()];
-    structRemappings = freeStructRemappings_;
-
-    // Only constants generated by `newToDeploy` exist at this stage
-    const constants = node.vVariables.flatMap((v) => {
-      assert(v.vValue !== undefined, 'Constants cannot be unanssigned');
-      return [`// ${v.documentation}`, `const ${v.name} = ${writer.write(v.vValue)};`].join('\n');
-    });
-
-    const structs = [...freeStructs, ...node.vStructs, ...(mainContract?.vStructs || [])].map((v) =>
-      writer.write(v),
-    );
-
-    const importFunctions = node.vFunctions.filter(
-      (f): f is CairoImportFunctionDefinition => f instanceof CairoImportFunctionDefinition,
-    );
-    const generatedFunctions = node.vFunctions.filter(
-      (f): f is CairoGeneratedFunctionDefinition => f instanceof CairoGeneratedFunctionDefinition,
-    );
-    const functions = node.vFunctions.filter(
-      (f) =>
-        !(f instanceof CairoGeneratedFunctionDefinition) &&
-        !(f instanceof CairoImportFunctionDefinition),
-    );
-
-    const writtenImportFuncs = getGroupedImports(
-      importFunctions
-        .sort((funcA, funcB) =>
-          `${funcA.path}.${funcA.name}`.localeCompare(`${funcB.path}.${funcB.name}`),
-        )
-        .filter((func, index, importFuncs) => func.name !== importFuncs[index - 1]?.name),
-    ).reduce((writtenImports, importFunc) => `${writtenImports}\n${importFunc}`, '');
-
-    const writtenGeneratedFuncs = generatedFunctions
-      .sort((funcA, funcB) => funcA.name.localeCompare(funcB.name))
-      .sort((funcA, funcB) => {
-        const stubA = funcA.functionStubKind;
-        const stubB = funcB.functionStubKind;
-        if (stubA === stubB) return 0;
-        if (stubA === FunctionStubKind.StructDefStub) return -1;
-        if (stubA === FunctionStubKind.StorageDefStub) return -1;
-        return 1;
-      })
-      .filter((func, index, genFuncs) => func.name !== genFuncs[index - 1]?.name)
-      .map((func) => writer.write(func));
-
-    const writtenFuncs = functions.map((func) => writer.write(func));
+    structRemappings = new Map();
 
     const contracts = node.vContracts.map((v) => writer.write(v));
 
-    return [
-      removeExcessNewlines(
-        [
-          '%lang starknet',
-          writtenImportFuncs,
-          ...constants,
-          ...structs,
-          ...writtenGeneratedFuncs,
-          ...writtenFuncs,
-          ...contracts,
-        ].join('\n\n\n'),
-        3,
-      ),
-    ];
+    return [removeExcessNewlines([...contracts].join('\n\n\n'), 3)];
   }
 
   private generateInterfaceNameMappings(node: SourceUnit) {
@@ -120,17 +52,4 @@ export class SourceUnitWriter extends CairoASTNodeWriter {
 
     interfaceNameMappings.set(node, map);
   }
-}
-
-function getGroupedImports(imports: CairoImportFunctionDefinition[]): string[] {
-  const processedImports: string[] = [];
-  imports.reduce((functionNames: string[], importNode, index) => {
-    functionNames.push(importNode.name);
-    if (importNode.path !== imports[index + 1]?.path) {
-      processedImports.push(`from ${importNode.path} import ${functionNames.join(', ')}`);
-      functionNames = [];
-    }
-    return functionNames;
-  }, []);
-  return processedImports;
 }
